restructuratuion de srcs/Profile for split different onglets in differents files. de meme for dir style/ and begin of onglet component
