+++ front/src/components/Profile/History.tsx
@@ -1,0 +1,9 @@
+
+export function History() {
+
+    return (
+        <div>
+
+        </div>
+    )
+}
