+++ front/src/index.tsx
@@ -1,12 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
+import './styles/index.scss';
 import reportWebVitals from './reportWebVitals';
 import App from './App';
 
 import { store } from './redux/store';
 import { Provider } from 'react-redux';
 import { CookiesProvider } from 'react-cookie';
+
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
 );
