+++ front/src/styles/index.scss
@@ -1,0 +1,12 @@
+body {
+    margin: 0;
+}
+
+@import "pages/chat.scss";
+@import "pages/connection.scss";
+@import "pages/game.scss";
+@import "pages/profile.scss";
+
+
+@import "components/nav.scss";
+@import "./styles.scss";
