+++ front/src/styles/pages/chat.scss
