+++ front/src/styles/pages/connection.scss
