+++ front/src/styles/pages/game.scss
